refactor global util functions
