+++ app/features/multiple_choice_quiz_generator/core.py
@@ -1,4 +1,4 @@
-from app.features.multiple_choice_quiz_generator.document_loaders import get_docs
+from app.utils.document_loaders import get_docs
 from app.services.logger import setup_logger
 from app.features.multiple_choice_quiz_generator.tools import QuizBuilder
 from app.api.error_utilities import LoaderError, ToolExecutorError
